// src/types/partials/ImagingStudySeries.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TId } from '../simpleTypes/Id';
import { TUnsignedInt } from '../simpleTypes/UnsignedInt';
import { TCoding } from '../partials/Coding';
import { TReference } from '../partials/Reference';
import { TDateTime } from '../simpleTypes/DateTime';
import { TImagingStudyPerformer } from '../partials/ImagingStudyPerformer';
import { TImagingStudyInstance } from '../partials/ImagingStudyInstance';

export type TImagingStudySeries = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    uid: TId;
    number?: TUnsignedInt;
    modality: TCoding;
    description?: String;
    numberOfInstances?: TUnsignedInt;
    endpoint?: TReference[];
    bodySite?: TCoding;
    laterality?: TCoding;
    specimen?: TReference[];
    started?: TDateTime;
    performer?: TImagingStudyPerformer[];
    instance?: TImagingStudyInstance[];
};